update readme with latest example script
