+++ requirements.txt
@@ -3,6 +3,7 @@ dataclasses>=0.6
 dataclasses-json>=0.6.6
 botocore>=1.34.70
 aiobotocore>=2.13.0
+requests>=2.32.2
 
 types-botocore>=1.0.2
 types-aiobotocore>=2.13.0
